Set PIN-PUK association for cards that don't have it set
sc_pkcs15_unblock_pin() in libopensc/pkcs15-pin.c wants to associate PIN
to be unblocked with its PUK to check, for example, whether provided PUK
conforms to its policy.

When this function is not able to find a relevant PUK is uses policy for
PIN to be unblocked instead to check provided PUK which causes problems if
PIN and PUK policies differ.

Set PIN-PUK association for cards where it was unset and where this
association was either obvious or described in code or specs.

This leaves SmartCard-HSM for which the same change should possibly be
done, but wasn't since I don't have enough information about it.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>