PTEID changes for 2048 bit keys and SHA256 hash
PtEid card can sign with  1024 or 2048bit RSA, and SHA1 or SHA256 hashes
as long as the data to be signed is less the 36 bytes.

v3 cards can accept a SHA256 hash without the digest header
and add the header back on before padding and signing.

This commit depends on  "Improved handling of Hash"

 Date:      Fri Feb 3 08:47:16 2017 -0600
 Changes to be committed:
	modified:   src/libopensc/card-gemsafeV1.c