Improved Handling of Hashes listing hashes acceptable to card
Some cards have limitations on what hashes they can support. A card may need to
add the digest for a hash or may have a limit on the size of the hash and digest header
or it may need to examine the digest header before padding.

The limitation is independent of where the hash and digest were created. A calling
application may have create the hash and digest and then used PKCS#11 CKM_RSA_PKCS
to pad and sign. Or the calling application uses mechanism CKM_SHA_RSA_PKCS requesting
OpenSC in software or hardware to create the hash and digest.

These limitations showxn up with the PtEid card that can sign a SHA256 hash
and digest but only if the hash without the digest is sent to the card. Other cards
may have similar problem.

This commit adds two flags.

The first flag and additional data is added to sc_security_env_t:
	unsigned long algorithm_flags_original;
	const u8 *in_data;
	size_t in_data;

sc_pkcs15_compute_signature copies the algorithm_flags to algorithm_flags_original
before it changes them so a card's set_security_env can look at at both the original
and modified algorithm_flags. It also copies the pointer data and datalen.
(Cards can ignore these.)

The second flag is added to to sc_card_t:
	unsigned long  restrict_to_these_hashes

pkcs11/framework-pkcs15.c and openssl.c are overly aggressive in trying to create PKCS#11
mechanisms for every hash OpenSSL supports even if the card could never sign that hash.
The restrict_to_these_hashes (defaults to zero meaning no restrictions) is a list of
all the hashes that the drive's _init routine knows the card can handle.

Thus a calling application will not be mislead into using a mechanism the card cannot
support.

For example, the PtEid card can not handle CKM_SHA224_RSA_PKCS, CKM_SHA384_RSA_PKCS
or CKM_SHA512_RSA_PKCS or can not handle one of these if the application created the hash
and uses RSA_PKCS.

For example for PtEid gemsafe_init sets:

	/* Card has 36 byte limit on size of data to pad. */
	card->restrict_to_these_hashes  = SC_ALGORITHM_RSA_HASH_SHA1;
	card->restrict_to_these_hashes |= SC_ALGORITHM_RSA_HASH_MD5;
	card->restrict_to_these_hashes |= SC_ALGORITHM_RSA_HASH_MD5_SHA1;
	card->restrict_to_these_hashes |= SC_ALGORITHM_RSA_HASH_RIPEMD160;

	/* GemSAFE V3 cards support SHA256 */
	/* by sending only hash and flag to tell card add the SHA256 prefix */

	if (card->type == SC_CARD_TYPE_GEMSAFEV1_PTEID ||
		card->type == SC_CARD_TYPE_GEMSAFEV1_SEEID)
			card->restrict_to_these_hashes |= SC_ALGORITHM_RSA_HASH_SHA256;

Code for card-gemsafeV1.c is in a seperate commit.

 Date:      Mon Feb 27 08:13:00 2017 -0600
 Changes to be committed:
	modified:   src/libopensc/opensc.h
	modified:   src/libopensc/pkcs15-sec.c
	modified:   src/pkcs11/framework-pkcs15.c
	modified:   src/pkcs11/openssl.c