Improved Handling of Hashes
Some cards have restrictions on what hashes they can support. The card may need to
add the digest for a hash, or may limit the size of the hash and digest header
or it may examine the digest header before padding.

 The restriction is independent of where the hash and digest were created. A calling
 PKCS#11 application may have create the hash and digest and used mechanism CKM_RSA_PKCS
 to pad and sign. Or use mechanism CKM_SHA_RSA_PKCS requesting OpenSC in software or
 hardware to create the hash and digest.

 These restrictions have shown up with the PtEid card that can sign a SHA256 hash
 and digest but only if the hash without the digest is sent to the card. Other cards
 may have similar problem.

 This commit adds two flags. The first flag is added to sc_security_env_t:
     unsigned long algorithm_hash;

sc_pkcs15_compute_signature looks at its flags and at the data to be signed to set
senv.algorithm_hash The sc_pkcs1_strip_digest_info_prefix is used to check if the
data looks like a known hash with a digest created externally by the calling
application or a hash and digest created internally by OpenSC.
The card drivers set_security_env can then use this information.

The second flag is added to to sc_card_t:
     unsigned long  negate_hashes;

pkcs11/framework-pkcs15.c and openssl.c are overly aggressive in trying to create PKCS#11
mechanisms for every hash OpenSSL supports even if the card could never sign that hash.
The negate_hashes (defaults to zero) is a list of known hashes that the driver's _init
routine knows the card can not handle.

Thus a calling application will not be mislead into using a mechanism the card cannot
support.

For example, the PtEid card can not handle CKM_SHA224_RSA_PKCS, CKM_SHA384_RSA_PKCS
or CKM_SHA512_RSA_PKCS or can not handle one of these if the application created the hash
and uses RSA_PKCS.

For example for PtEid gemsafe_init sets:

      card->negate_hashes = SC_ALGORITHM_RSA_HASH_SHA224 |
            SC_ALGORITHM_RSA_HASH_SHA384  |
            SC_ALGORITHM_RSA_HASH_SHA512;

Code for card-gemsafeV1.c is in a seperate commit.

 Date:      Tue Feb 14 10:21:36 2017 -0600
 Changes to be committed:
	modified:   src/libopensc/opensc.h
	modified:   src/libopensc/pkcs15-sec.c
	modified:   src/pkcs11/framework-pkcs15.c
	modified:   src/pkcs11/openssl.c