iso7816: disable SM for GET RESPONSE
The result of GET RESPONSE contains a potentially fragmented response
that is protected with SM. We disable SM here to let the upper layers
assemble and verify the complete SM response.

ISO 7816 is not very clear about usage of GET RESPONSE in SM, but we
have good evidence that it's indended to be read the way we implemented
it, see https://github.com/OpenSC/OpenSC/issues/970.