PKCS#11 does not define a CKA_VALUE for public keys
OpenSC opennssl.c  in sc_pkcs11_verify_data assumes that it can
retieve the CKA_VALUE  for a public key object, and expect it to
be usable as RSA.

But internally  sc_pkcs15_pubkey can have a "direct" or "spki"
version of the public key as defined by PKCS#15.  card drivers
or pkcs15-<card> routines may store either the "direct" or "spki"
versions. A get attribute request for CKA_VALUE  for a public key
will return either the direct, spki ora derived rsa verison of the
pubkey.

This commit will use d2i_PUBKEY which takes a spki version and
returns an EVP_KEY. If it fails, the current method,
d21_PublicKey(EVP_PKEY_RSA,...) is tried to get an EVP_KEY
which only works if it is RSA.

The problem was found while testing pkcs11-tool -t -l  where
the  verify tests would fail with a CKR_GENERAL_ERROR because
the card driver stored the public key as a spki.

Please enter the commit message for your changes. Lines starting
 On branch verify-pubkey-as-spki
 Changes to be committed:
	modified:   pkcs11/openssl.c