sc_pkcs15_read_der_file fails when used with piv to read pubkey from file
sc_pkcs15_read_der_file reads 16 bytes to get the tag. It then calls
  sc_asn1_read_tag(&body, len, &cla_out, &tag_out, &bodylen);
with len = 16 passed into sc_asn1_read_tag as buflen.

sc_asn1_read_tag pasers the tag then checks to make sure the data
does overflow the size left.

At this point len = the size of the data  for the tag, which in most cases
is greater the 16.
  if (len > left)
    return SC_ERROR_INVALID_ASN1_OBJECT;

This causes sc_pkcs15_read_der_file to fail.

The change replaces the call with:
  sc_asn1_read_tag(&body, 0xffff, &cla_out, &tag_out, &bodylen);
Thus the if above does not fail.

The PIV code  used sc_asn1_read_tag  with 0xffff to get the size of
an object so to can allocate a buffer then read the full object.

As far as I can tell the only place this ie ever used (and used to work)
in from pkcs15-piv.c  when there is key on the card but no certificate.
and is used when trying to sign a certificate request to get the public key.

 Changes to be committed:
	modified:   pkcs15-pubkey.c