opensc-explorer: accept more than two arguments in 'do_put'
Generate the input to the 'do_put' command from a series of
hex sequences and strings instead limiting it to only one.
While doing so, increase the buffer to allow for more data
to be written to DOs.

This allows for efficient input of data while keeping universality:
most data as strings, only non-printable ones as hex.