Use sc_asn1_read_tag to read first tag of partially block
Sc_asn1_read_tag can return SC_ERROR_ASN1_END_OF_CONTENTS
which indicates the tag and length are OK, but any value
is not completely contained in the buffer supplied. card-piv.c
can use this when reading just the beginning of a object to
determine the size of a buffer needed to hold the object.

 On branch piv-buflen-issue
 Changes to be committed:
	modified:   card-piv.c

 interactive rebase in progress; onto fcd719d
 Last commands done (2 commands done):
    pick 95e5263 Fix card-piv.c use of sc_asn1_read_tag
    squash d86604f Use sc_asn1_read_tag to read first tag of partially block
 Changes to be committed:
	modified:   src/libopensc/card-piv.c