PIV implemention of dynamic response buffer using BER_TLV in response
The PIV card uses a number of ISO 7816-4 instructions where the payload shall be encoded in BER-TLV.
ISO 7816-4 recommends these be DER encoded, which they in the PIV case. These include the odd numbered
GET DATA, GENERAL AUTHENTICATE and GENERATE ASYMMETRIC KEY PAIR. Thus when reading any of these payloads,
the size of the data to read is defined by the BER_TLV tag plus the size of the tag. The only way to
determine the size is to read the object.

Current PIV code does this in 2 steps. It uses GET DATA to read the first 8 bytes that will contain the
BER_TLV tag. Then it allocates a buffer then does a second GET DATA to read the tag and data.

In some recent fuzzing tests the first tag did not match the second tag causing problems.

This commit now allocates a starting buffer, reads the first 256 bytes, parses the tag, checks its length,
reallocates the buffer (if needed) then continues to read the rest of the data using GET RESPONSE. This
is done by using the apdu SC_APDU_FLAGS_NO_GET_RESP to get control back after the initial 256 bytes of
data are read for T1 and for T0 after the first GET RESPONSE command completes.

Placing this code in to apdu.c could be possible but very difficult. So this code is completely contained
in card-piv.c

 On branch piv-dynamic-resp

 Changes to be committed:
	modified:   src/libopensc/card-piv.c