Improved Handling of CK_SLOT_ID values returned to C_GetSlotList
Previously the CK_SLOT_IDs where derived from the place on the
virtual_slot list. This means that if a reader was removed, the position
on the list could change. An application such as Firefox is also keeping
track of data based on the previous slot list. So the application
could get confused.

PKCS11 v2.30 code in in OpenSC to keep deleted slots on the list until the
application has see the slot once since itmiht have been deleted. But the
slot also looked like an available hotplug which might g reused before
the application ha seen it. This was fixed so it could not be reused until
the application had a chance to see it.

The CK_SLOT_ID now starts at 0, and increased by 1 each time a new slot
is created which occues when a reader is inserted. There is an upper limit
of 2^32 calls to create_slot. In practics this will never be reached but
the code has a "TODO" to add code to reuse slot IDs.

Pkcs11-tool.c has improved wording for listing slots.
 On branch PKCS11-SLOTS
 Changes to be committed:
	modified:   pkcs11-global.c
	modified:   sc-pkcs11.h
	modified:   slot.c
	modified:   ../tools/pkcs11-tool.c