Add debugging for PKCS11 locking in threaded environment
this is test code to help debut issue #1966
it is being submitted as a PR to get the windows artifacts
so user can test it.

It will show what is passed to C_Initialize by Java and
show enter and leave log messages for sc_pkcs11_lock and sc_pkcs11_unlock
if thay are actualling using teh mutex.

 On branch PKCS11-lock
 Changes to be committed:
	modified:   pkcs11-global.c