YubiKey5 PIV changes using Attestation
Yubico version 4 and 5 devices allow a pin and touch policy to be set
during generation or importation of a key. They also support the creation
of an attestation certificate for a key with certificate extensions, one
of which has the pin and touch policies for the key.

NIST 800-73 PIV specifications says the 9C key will require the
PIN for every use of the key. PIV cards enforce this "PIN Always" policy
on the card. (the 9E key does not require a PIN.)

But the Yubico devices have extensions that do not always follow the
NIST specifications. This has lead to caching of PINs and changes
to opensc.conf to be able to "turn off" the "PIN Always" policy.

If there is an attestation certificate for the 9C PIV Signing key,
and the pin policy extension is not the default=0, or "always=3" this
means the card will not enforce the NIST PIV "PIN Always".
card-piv.c and pkcs15-piv.c will then mark the 9C key as not
requiring "PIN Always" aka "user_consent" aka "CKA_ALWAYS_AUTHENTICATE"
and the key will be treated as any other key.

This PR could have gone further, and checked and set pin "NEVER"
much like the PIV 9E key and touch polices, but is not clear how to
handle via PKCS#11 how to tell user to touch the button with or without
a PIN.

Tested with Yubikey 5 NFC version 5.2.6.

The 5.2.6 device appears to have fixed all the other "card_issues"
as listed in card-piv.c for other cards and keeps the login state correctly.

See:
https://developers.yubico.com/PIV/Introduction/PIV_attestation.html
https://developers.yubico.com/yubico-piv-tool/Manuals/yubico-piv-tool.1.html
Or contact Yubico support.

 On branch YubiKey5
 Changes to be committed:
	modified:   card-piv.c
	modified:   cardctl.h
	modified:   cards.h
	modified:   pkcs15-piv.c