Yubikey touch policy hint to user added to pin prompt
Yubikey 4.3 and higher devices with an attestion certificate
are used to get the pin and touch policies.

Older devices may have a touch policy, but there is no way to
test for this in the code.

Tokens blink at a 1 second rate. If not touched for some time
 looks like a minute, the operation is aborted.

if one or more keys have touch policy in any attestation certificate
requiring touching, a message is added to the pin prompt.
Currently ",Yubikey has touch policy" is added.

A large section of code to add pins was move to after the private
keys are added in order to add the prompt.

 On branch YubiKey5
 Changes to be committed:
	modified:   src/libopensc/pkcs15-piv.c