Travis.yml - Improve PIV testing
Fix Issue #2094

Current test using PivApplet/bin/:jcardsim/target/jcardsim-3.0.5
only creates 2 private keys. These are never seen by a PIV client,
as the way private keys are found is via a certificate which contains
the public key (SPKI) which contains the type of key, RSA or ECC
and the size or curve OID.

This  change creates the public key and self signed certificate
for the two keys so  pkcs11-tool can list and test the crypto.

 Please enter the commit message for your changes. Lines starting
 On branch PIV-4-extensions
 Changes to be committed:
	modified:   .travis.yml